requirements.txt
Django==4.2.7
Pillow==9.5.0
djangorestframework==3.14.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.1
structlog==24.1.0 